chore: move commands from admin back to strapi/strapi
