+++ packages/core/admin/admin/src/index.ts
@@ -1,4 +1,3 @@
-export * from './components/DefaultDocument';
 export * from './render';
 
 export type { Store } from './core/store/configure';
