+++ packages/core/admin/package.json
@@ -20,12 +20,6 @@
     }
   ],
   "exports": {
-    ".": {
-      "types": "./index.d.ts",
-      "source": "./index.js",
-      "require": "./index.js",
-      "default": "./index.js"
-    },
     "./strapi-admin": {
       "types": "./dist/admin/src/index.d.ts",
       "source": "./admin/src/index.ts",
@@ -33,25 +27,24 @@
       "require": "./dist/admin/index.js",
       "default": "./dist/admin/index.js"
     },
+    "./_internal": {
+      "types": "./dist/_internal/index.d.ts",
+      "source": "./_internal/index.ts",
+      "import": "./dist/_internal.mjs",
+      "require": "./dist/_internal.js",
+      "default": "./dist/_internal.js"
+    },
     "./strapi-server": {
       "types": "./dist/server/src/index.d.ts",
       "source": "./server/src/index.js",
       "require": "./strapi-server.js",
       "default": "./strapi-server.js"
     },
-    "./cli": {
-      "source": "./_internal/index.ts",
-      "import": "./dist/cli.mjs",
-      "require": "./dist/cli.js",
-      "default": "./dist/cli.js"
-    },
     "./package.json": "./package.json"
   },
   "files": [
     "dist/",
-    "strapi-server.js",
-    "index.js",
-    "index.d.ts"
+    "strapi-server.js"
   ],
   "scripts": {
     "build": "pack-up build",
@@ -69,7 +62,6 @@
   },
   "dependencies": {
     "@casl/ability": "6.5.0",
-    "@pmmmwh/react-refresh-webpack-plugin": "0.5.10",
     "@radix-ui/react-context": "1.0.1",
     "@radix-ui/react-toolbar": "1.0.4",
     "@reduxjs/toolkit": "1.9.7",
@@ -85,28 +77,17 @@
     "axios": "1.6.0",
     "bcryptjs": "2.4.3",
     "boxen": "5.1.2",
-    "browserslist": "^4.22.1",
-    "browserslist-to-esbuild": "1.2.0",
     "chalk": "^4.1.2",
-    "chokidar": "3.5.3",
     "codemirror5": "npm:codemirror@^5.65.11",
     "cross-env": "^7.0.3",
-    "css-loader": "^6.8.1",
     "date-fns": "2.30.0",
-    "dotenv": "14.2.0",
-    "esbuild": "0.19.2",
-    "esbuild-loader": "^2.21.0",
-    "esbuild-register": "3.5.0",
     "execa": "5.1.1",
     "fast-deep-equal": "3.1.3",
-    "find-root": "1.1.0",
-    "fork-ts-checker-webpack-plugin": "8.0.0",
     "formik": "2.4.0",
     "fractional-indexing": "3.2.0",
     "fs-extra": "10.0.0",
     "highlight.js": "^10.4.1",
     "history": "^4.9.0",
-    "html-webpack-plugin": "5.5.0",
     "immer": "9.0.19",
     "inquirer": "8.2.5",
     "invariant": "^2.2.4",
@@ -129,14 +110,11 @@
     "markdown-it-mark": "^3.0.1",
     "markdown-it-sub": "^1.0.0",
     "markdown-it-sup": "1.0.0",
-    "mini-css-extract-plugin": "2.7.6",
     "node-schedule": "2.1.0",
     "ora": "5.4.1",
-    "outdent": "0.8.0",
     "p-map": "4.0.0",
     "passport-local": "1.0.0",
     "pluralize": "8.0.0",
-    "prettier": "2.8.4",
     "prop-types": "^15.8.1",
     "qs": "6.11.1",
     "react": "^18.2.0",
@@ -153,8 +131,6 @@
     "react-router-dom": "5.3.4",
     "react-select": "5.7.0",
     "react-window": "1.8.8",
-    "read-pkg-up": "7.0.1",
-    "resolve-from": "5.0.0",
     "rimraf": "3.0.2",
     "sanitize-html": "2.11.0",
     "semver": "7.5.4",
@@ -162,20 +138,14 @@
     "slate": "0.94.1",
     "slate-history": "0.93.0",
     "slate-react": "0.98.3",
-    "style-loader": "3.3.1",
     "styled-components": "5.3.3",
     "typescript": "5.3.2",
-    "webpack": "^5.88.1",
-    "webpack-bundle-analyzer": "^4.9.0",
-    "webpack-dev-middleware": "6.1.1",
-    "webpack-hot-middleware": "2.25.4",
     "yup": "0.32.9"
   },
   "devDependencies": {
     "@strapi/admin-test-utils": "4.15.5",
     "@strapi/pack-up": "4.15.5",
     "@strapi/plugin-content-manager": "4.15.5",
-    "@strapi/strapi": "4.15.5",
     "@testing-library/dom": "9.2.0",
     "@testing-library/react": "14.0.0",
     "@testing-library/user-event": "14.4.3",
@@ -203,7 +173,6 @@
   },
   "peerDependencies": {
     "@strapi/data-transfer": "4.15.5",
-    "@strapi/strapi": "^4.3.4",
     "react": "^17.0.0 || ^18.0.0",
     "react-dom": "^17.0.0 || ^18.0.0",
     "react-router-dom": "^5.2.0",
