+++ packages/core/strapi/package.json
@@ -103,6 +103,7 @@
     "watch": "pack-up watch"
   },
   "dependencies": {
+    "@pmmmwh/react-refresh-webpack-plugin": "0.5.10",
     "@strapi/admin": "4.15.5",
     "@strapi/core": "4.15.5",
     "@strapi/data-transfer": "4.15.5",
@@ -120,17 +121,41 @@
     "@strapi/typescript-utils": "4.15.5",
     "@strapi/utils": "4.15.5",
     "boxen": "5.1.2",
+    "browserslist": "^4.22.1",
+    "browserslist-to-esbuild": "1.2.0",
     "chalk": "4.1.2",
+    "chokidar": "3.5.3",
     "cli-table3": "0.6.2",
     "commander": "8.3.0",
     "copyfiles": "2.4.1",
+    "css-loader": "^6.8.1",
+    "dotenv": "14.2.0",
+    "esbuild": "0.19.2",
+    "esbuild-loader": "^2.21.0",
+    "esbuild-register": "3.5.0",
     "execa": "5.1.1",
+    "find-root": "1.1.0",
+    "fork-ts-checker-webpack-plugin": "8.0.0",
     "fs-extra": "10.0.0",
+    "html-webpack-plugin": "5.5.0",
     "inquirer": "8.2.5",
     "lodash": "4.17.21",
+    "mini-css-extract-plugin": "2.7.6",
     "ora": "5.4.1",
+    "outdent": "0.8.0",
     "pkg-up": "3.1.0",
+    "prettier": "2.8.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "read-pkg-up": "7.0.1",
+    "resolve-from": "5.0.0",
+    "semver": "7.5.4",
+    "style-loader": "3.3.1",
     "typescript": "5.3.2",
+    "webpack": "^5.88.1",
+    "webpack-bundle-analyzer": "^4.9.0",
+    "webpack-dev-middleware": "6.1.1",
+    "webpack-hot-middleware": "2.25.4",
     "yup": "0.32.9"
   },
   "devDependencies": {
@@ -140,6 +165,10 @@
     "@types/node": "18.18.4",
     "eslint-config-custom": "4.15.5",
     "tsconfig": "4.15.5"
+  },
+  "peerDependencies": {
+    "react": "^17.0.0 || ^18.0.0",
+    "react-dom": "^17.0.0 || ^18.0.0"
   },
   "engines": {
     "node": ">=18.0.0 <=20.x.x",
