--- packages/core/strapi/src/commands/actions/watch-admin/action.ts
@@ -1,14 +1,0 @@
-import { watchAdmin } from '@strapi/admin';
-import { CLIContext } from '../../types';
-
-interface WatchAdminOptions extends CLIContext {
-  browser: boolean;
-}
-
-export default async ({ browser, logger }: WatchAdminOptions) => {
-  logger.warn('[@strapi/strapi]: watch-admin is deprecated, please use strapi develop instead');
-
-  await watchAdmin({
-    browser,
-  });
-};
