+++ packages/core/strapi/src/commands/index.ts
@@ -21,7 +21,8 @@ import enableTelemetry from './actions/telemetry/enable/command';
 import generateTemplates from './actions/templates/generate/command';
 import generateTsTypes from './actions/ts/generate-types/command';
 import versionCommand from './actions/version/command';
-import watchAdminCommand from './actions/watch-admin/command';
+import buildCommand from './actions/build-command/command';
+import developCommand from './actions/develop/command';
 
 import buildPluginCommand from './actions/plugin/build-command/command';
 import watchPluginCommand from './actions/plugin/watch/command';
@@ -52,7 +53,8 @@ const strapiCommands = [
   generateTemplates,
   generateTsTypes,
   versionCommand,
-  watchAdminCommand,
+  buildCommand,
+  developCommand,
   /**
    * Plugins
    */
@@ -66,10 +68,6 @@ const buildStrapiCommand = async (argv: string[], command = new Command()) => {
     // eslint-disable-next-line @typescript-eslint/no-var-requires
     const dtsCommands = require(require.resolve('@strapi/data-transfer')).commands;
     strapiCommands.push(...dtsCommands);
-
-    // eslint-disable-next-line @typescript-eslint/no-var-requires, node/no-missing-require
-    const adminCommands = require(require.resolve('@strapi/admin/cli')).commands;
-    strapiCommands.push(...adminCommands);
   } catch (e) {
     // noop
   }
