+++ packages/core/strapi/src/node/core/errors.ts
