+++ packages/core/strapi/src/node/core/monorepo.ts
