+++ packages/core/strapi/src/node/webpack/aliases.ts
