+++ packages/core/strapi/src/node/webpack/build.ts
@@ -3,15 +3,15 @@ import webpack from 'webpack';
 import { mergeConfigWithUserConfig, resolveProductionConfig } from './config';
 import { isError } from '../core/errors';
 
-import type { BuildContext } from '../createBuildContext';
+import type { BuildContext } from '../create-build-context';
 
-const build = async (ctx: BuildContext) =>
-  new Promise(async (resolve, reject) => {
-    const config = await resolveProductionConfig(ctx);
-    const finalConfig = await mergeConfigWithUserConfig(config, ctx);
+const build = async (ctx: BuildContext) => {
+  const config = await resolveProductionConfig(ctx);
+  const finalConfig = await mergeConfigWithUserConfig(config, ctx);
 
-    ctx.logger.debug('Webpack config', finalConfig);
+  ctx.logger.debug('Webpack config', finalConfig);
 
+  return new Promise((resolve, reject) => {
     webpack(finalConfig, (err, stats) => {
       if (stats) {
         if (stats.hasErrors()) {
@@ -22,7 +22,7 @@ const build = async (ctx: BuildContext) =>
             })
           );
 
-          reject(false);
+          reject();
         } else if (ctx.options.stats) {
           ctx.logger.info(
             stats.toString({
@@ -37,9 +37,10 @@ const build = async (ctx: BuildContext) =>
 
       if (err && isError(err)) {
         ctx.logger.error(err.message);
-        reject(false);
+        reject();
       }
     });
   });
+};
 
 export { build };
