+++ packages/core/strapi/tsconfig.json
@@ -1,7 +1,8 @@
 {
   "extends": "tsconfig/base.json",
   "compilerOptions": {
-    "noEmit": true
+    "noEmit": true,
+    "rootDir": "./src"
   },
   "include": ["src"],
   "exclude": ["node_modules"]
